+++ src/ofApp.h
@@ -36,12 +36,12 @@ public:
     void gotMessage(ofMessage msg);
     void newHapMovie(std::string path, int index, ofVec3f* initPts, int width, int height, ofxYAML& config, int videoIndex);
     void displayIncomingData(int width, int height);
-    void onsetOccured(int width, int height, int frame_num);
-    void drawScreen(int width, int height, int frameNum, bool isNRT);
-    void setValsFromCSV(int width, int height, vector<float>& csv_data, int frame_num);
+    void onsetOccured(int width, int height, unsigned long long frame_num);
+    void drawScreen(int width, int height, unsigned long long frameNum, bool isNRT);
+    void setValsFromCSV(int width, int height, vector<float>& csv_data, unsigned long long frame_num);
     void incrementVecHistoryCounter();
-    void processReaperMarker(string& cmd,int width, int height, int frame_num);
-    void setActiveIndices(int* ai,int width, int height, int frame_num);
+    void processReaperMarker(string& cmd,int width, int height, unsigned long long frame_num);
+    void setActiveIndices(int* ai,int width, int height, unsigned long long frame_num);
     void prUpdate(bool isNRT);
     void drawFluidFlow(int width, int height);
     void runNrtRender(int width,int height);
